+++ boards/nxp/imx95_evk/imx95_evk_mimx9596_m7_ddr_defconfig
@@ -17,6 +17,7 @@ CONFIG_CONSOLE=y
 CONFIG_MBOX=y
 CONFIG_MBOX_INIT_PRIORITY=0
 CONFIG_ARM_SCMI=y
+CONFIG_ARM_SCMI_NXP_VENDOR_EXTENSIONS=y
 
 # kernel-related configurations
 CONFIG_XIP=n
