+++ boards/nxp/imx95_evk/imx95_evk_mimx9596_m7_defconfig
@@ -13,3 +13,4 @@ CONFIG_XIP=y
 CONFIG_MBOX=y
 CONFIG_MBOX_INIT_PRIORITY=0
 CONFIG_ARM_SCMI=y
+CONFIG_ARM_SCMI_NXP_VENDOR_EXTENSIONS=y
