+++ drivers/firmware/scmi/nxp/Kconfig
@@ -1,0 +1,8 @@
+# Copyright 2025 NXP
+# SPDX-License-Identifier: Apache-2.0
+
+config ARM_SCMI_NXP_VENDOR_EXTENSIONS
+	bool "Allow NXP specific SCMI features"
+	select CRC
+	help
+	  When enabled, additional SCMI features specific to NXP will be available
