+++ include/zephyr/drivers/firmware/scmi/shmem.h
@@ -19,6 +19,15 @@
 #define SCMI_SHMEM_CHAN_STATUS_BUSY_BIT BIT(0)
 #define SCMI_SHMEM_CHAN_FLAG_IRQ_BIT BIT(0)
 
+struct scmi_shmem_layout {
+	volatile uint32_t res0;
+	volatile uint32_t chan_status;
+	volatile uint32_t res1[2];
+	volatile uint32_t chan_flags;
+	volatile uint32_t len;
+	volatile uint32_t msg_hdr;
+};
+
 struct scmi_message;
 
 /**
@@ -63,5 +72,25 @@ void scmi_shmem_update_flags(const struct device *shmem,
  * @param shmem pointer to shmem device
  */
 uint32_t scmi_shmem_channel_status(const struct device *shmem);
+
+/**
+ * @brief Process vendor specific features when writing message
+ *
+ * @param layout layout of the message
+ *
+ * @retval 0 if successful
+ * @retval negative errno if failure
+ */
+int scmi_shmem_vendor_write_message(struct scmi_shmem_layout *layout);
+
+/**
+ * @brief Process vendor specific features when reading message
+ *
+ * @param layout layout of the message
+ *
+ * @retval 0 if successful
+ * @retval negative errno if failure
+ */
+int scmi_shmem_vendor_read_message(const struct scmi_shmem_layout *layout);
 
 #endif /* _INCLUDE_ZEPHYR_DRIVERS_FIRMWARE_SCMI_SHMEM_H_ */
